cambios

+++ ecommerce_apis/apps/producto/models.py
@@ -6,4 +6,4 @@ class Producto(models.Model):
 
     nombre = models.CharField(max_length=100)
     precio = models.FloatField()
-    stock = models.DecimalField(max_digits=8, decimal_places=2)
+    stock = models.IntegerField()
